Renaming and cleaning up tests Part 1
