+++ bundles/net.tourbook.tests/src/test/export/files/LongsPeak-CamouflageSpeed-15kmh-BikingActivity.tcx
